// packages/sdk/src/redux/constants.ts
export enum ReduxActionTypes {
  SetAccount = '@archanova/wallet-sdk/SetAccount',
  SetAccountBalance = '@archanova/wallet-sdk/SetAccountBalance',
  SetAccountDevice = '@archanova/wallet-sdk/SetAccountDevice',
  SetDeviceAddress = '@archanova/wallet-sdk/SetDeviceAddress',
  SetGasPrice = '@archanova/wallet-sdk/SetGasPrice',
  SetNetworkVersion = '@archanova/wallet-sdk/SetNetworkVersion',
  SetInitialized = '@archanova/wallet-sdk/SetInitialized',
  SetAuthenticated = '@archanova/wallet-sdk/SetAuthenticated',
  SetConnected = '@archanova/wallet-sdk/SetConnected',
  SetIncomingAction = '@archanova/wallet-sdk/SetIncomingAction',
}